griptape-observability/requirements.txt
python-dotenv==1.0.1
argparse==1.4.0
griptape[drivers-observability-griptape-cloud]==0.33.1